server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-genai
python-dotenv
